requirements.txt
anthropic>=0.39.0
feedparser>=6.0.0
PyYAML>=6.0.0
trafilatura>=1.6.0
notion-client>=2.0.0

# YouTube transcript (via RSS + youtube-transcript-api)
youtube-transcript-api>=0.6.0

# Twitter (Playwright browser automation)
playwright>=1.40.0

# Gmail API - optional
google-api-python-client>=2.100.0
google-auth>=2.23.0

# Email (IMAP) - optional
imapclient>=2.3.0
